made header encryption indicator 8-bit type

+++ header_encryption.h
@@ -23,8 +23,8 @@ uint32_t packet_header_decrypt (uint8_t packet[], uint8_t packet_len,
                                 char * community_name, he_context_t * ctx);
 
 
-int32_t packet_header_decrypt_if_required (uint8_t packet[], uint16_t packet_len,
-                                           struct sn_community * communities);
+int8_t packet_header_decrypt_if_required (uint8_t packet[], uint16_t packet_len,
+                                          struct sn_community * communities);
 
 
 int32_t packet_header_encrypt (uint8_t packet[], uint8_t header_len, he_context_t * ctx);
